backend/requirements.txt
numpy>=1.24.3
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
openai==1.12.0
pydantic==1.10.11
loguru==0.7.0
python-dotenv==1.0.0
